Add a UploadFilesComponent.
